♻️ refactor: refactor api key form locale
