+++ src/app/settings/llm/Bedrock/index.tsx
@@ -13,7 +13,7 @@ import { LLMProviderConfigKey } from '../const';
 const providerKey: GlobalLLMProviderKey = 'bedrock';
 
 const BedrockProvider = memo(() => {
-  const { t } = useTranslation('setting');
+  const { t } = useTranslation('modelProvider');
 
   return (
     <ProviderConfig
@@ -22,22 +22,22 @@ const BedrockProvider = memo(() => {
           children: (
             <Input.Password
               autoComplete={'new-password'}
-              placeholder={t(`llm.${providerKey}.accessKeyId.placeholder`)}
+              placeholder={t(`${providerKey}.accessKeyId.placeholder`)}
             />
           ),
-          desc: t(`llm.${providerKey}.accessKeyId.desc`),
-          label: t(`llm.${providerKey}.accessKeyId.title`),
+          desc: t(`${providerKey}.accessKeyId.desc`),
+          label: t(`${providerKey}.accessKeyId.title`),
           name: [LLMProviderConfigKey, providerKey, 'accessKeyId'],
         },
         {
           children: (
             <Input.Password
               autoComplete={'new-password'}
-              placeholder={t(`llm.${providerKey}.secretAccessKey.placeholder`)}
+              placeholder={t(`${providerKey}.secretAccessKey.placeholder`)}
             />
           ),
-          desc: t(`llm.${providerKey}.secretAccessKey.desc`),
-          label: t(`llm.${providerKey}.secretAccessKey.title`),
+          desc: t(`${providerKey}.secretAccessKey.desc`),
+          label: t(`${providerKey}.secretAccessKey.title`),
           name: [LLMProviderConfigKey, providerKey, 'secretAccessKey'],
         },
         {
@@ -51,8 +51,8 @@ const BedrockProvider = memo(() => {
               placeholder={'us-east-1'}
             />
           ),
-          desc: t(`llm.${providerKey}.region.desc`),
-          label: t(`llm.${providerKey}.region.title`),
+          desc: t(`${providerKey}.region.desc`),
+          label: t(`${providerKey}.region.title`),
           name: [LLMProviderConfigKey, providerKey, 'region'],
         },
       ]}
