--- src/features/Conversation/Error/APIKeyForm/Groq.tsx
@@ -1,60 +1,0 @@
-import { Groq } from '@lobehub/icons';
-import { Input } from 'antd';
-import { memo } from 'react';
-import { useTranslation } from 'react-i18next';
-
-import { ModelProvider } from '@/libs/agent-runtime';
-import { useGlobalStore } from '@/store/global';
-import { modelConfigSelectors } from '@/store/global/selectors';
-
-import { FormAction } from '../style';
-
-const GroqForm = memo(() => {
-  const { t } = useTranslation('error');
-  // const [showProxy, setShow] = useState(false);
-
-  const [apiKey, setConfig] = useGlobalStore((s) => [
-    modelConfigSelectors.groqAPIKey(s),
-    s.setModelProviderConfig,
-  ]);
-
-  return (
-    <FormAction
-      avatar={<Groq size={56} />}
-      description={t('unlock.apikey.Groq.description')}
-      title={t('unlock.apikey.Groq.title')}
-    >
-      <Input.Password
-        autoComplete={'new-password'}
-        onChange={(e) => {
-          setConfig(ModelProvider.Groq, { apiKey: e.target.value });
-        }}
-        placeholder={'*********************************'}
-        type={'block'}
-        value={apiKey}
-      />
-      {/*{showProxy ? (*/}
-      {/*  <Input*/}
-      {/*    onChange={(e) => {*/}
-      {/*      setConfig({ endpoint: e.target.value });*/}
-      {/*    }}*/}
-      {/*    placeholder={'https://api.openai.com/v1'}*/}
-      {/*    type={'block'}*/}
-      {/*    value={proxyUrl}*/}
-      {/*  />*/}
-      {/*) : (*/}
-      {/*  <Button*/}
-      {/*    icon={<Icon icon={Network} />}*/}
-      {/*    onClick={() => {*/}
-      {/*      setShow(true);*/}
-      {/*    }}*/}
-      {/*    type={'text'}*/}
-      {/*  >*/}
-      {/*    {t('unlock.apikey.addProxyUrl')}*/}
-      {/*  </Button>*/}
-      {/*)}*/}
-    </FormAction>
-  );
-});
-
-export default GroqForm;
