--- src/features/Conversation/Error/APIKeyForm/OpenRouter.tsx
@@ -1,40 +1,0 @@
-import { OpenRouter } from '@lobehub/icons';
-import { Input } from 'antd';
-import { memo } from 'react';
-import { useTranslation } from 'react-i18next';
-
-import { ModelProvider } from '@/libs/agent-runtime';
-import { useGlobalStore } from '@/store/global';
-import { modelConfigSelectors } from '@/store/global/selectors';
-
-import { FormAction } from '../style';
-
-const OpenRouterForm = memo(() => {
-  const { t } = useTranslation('error');
-  // const [showProxy, setShow] = useState(false);
-
-  const [apiKey, setConfig] = useGlobalStore((s) => [
-    modelConfigSelectors.openrouterAPIKey(s),
-    s.setModelProviderConfig,
-  ]);
-
-  return (
-    <FormAction
-      avatar={<OpenRouter size={56} />}
-      description={t('unlock.apikey.OpenRouter.description')}
-      title={t('unlock.apikey.OpenRouter.title')}
-    >
-      <Input.Password
-        autoComplete={'new-password'}
-        onChange={(e) => {
-          setConfig(ModelProvider.OpenRouter, { apiKey: e.target.value });
-        }}
-        placeholder={'*********************************'}
-        type={'block'}
-        value={apiKey}
-      />
-    </FormAction>
-  );
-});
-
-export default OpenRouterForm;
