--- src/features/Conversation/Error/APIKeyForm/TogetherAI.tsx
@@ -1,40 +1,0 @@
-import { Together } from '@lobehub/icons';
-import { Input } from 'antd';
-import { memo } from 'react';
-import { useTranslation } from 'react-i18next';
-
-import { ModelProvider } from '@/libs/agent-runtime';
-import { useGlobalStore } from '@/store/global';
-import { modelConfigSelectors } from '@/store/global/selectors';
-
-import { FormAction } from '../style';
-
-const TogetherAIForm = memo(() => {
-  const { t } = useTranslation('error');
-  // const [showProxy, setShow] = useState(false);
-
-  const [apiKey, setConfig] = useGlobalStore((s) => [
-    modelConfigSelectors.togetheraiAPIKey(s),
-    s.setModelProviderConfig,
-  ]);
-
-  return (
-    <FormAction
-      avatar={<Together size={56} />}
-      description={t('unlock.apikey.TogetherAI.description')}
-      title={t('unlock.apikey.TogetherAI.title')}
-    >
-      <Input.Password
-        autoComplete={'new-password'}
-        onChange={(e) => {
-          setConfig(ModelProvider.TogetherAI, { apiKey: e.target.value });
-        }}
-        placeholder={'*********************************'}
-        type={'block'}
-        value={apiKey}
-      />
-    </FormAction>
-  );
-});
-
-export default TogetherAIForm;
