+++ src/features/Conversation/Error/APIKeyForm/index.tsx
@@ -5,19 +5,11 @@ import { Center, Flexbox } from 'react-layout-kit';
 
 import { ModelProvider } from '@/libs/agent-runtime';
 import { useChatStore } from '@/store/chat';
+import { GlobalLLMProviderKey } from '@/types/settings';
 
-import AnthropicForm from './Anthropic';
 import BedrockForm from './Bedrock';
-import GoogleForm from './Google';
-import GroqForm from './Groq';
-import MistralForm from './Mistral';
-import MoonshotForm from './Moonshot';
-import OpenAIForm from './OpenAI';
-import OpenRouterForm from './OpenRouter';
-import PerplexityForm from './Perplexity';
-import TogetherAIForm from './TogetherAI';
-import ZeroOneForm from './ZeroOne';
-import ZhipuForm from './Zhipu';
+import ProviderApiKeyForm from './ProviderApiKeyForm';
+import ProviderAvatar from './ProviderAvatar';
 
 interface APIKeyFormProps {
   id: string;
@@ -29,62 +21,46 @@ const APIKeyForm = memo<APIKeyFormProps>(({ id, provider }) => {
 
   const [resend, deleteMessage] = useChatStore((s) => [s.internalResendMessage, s.deleteMessage]);
 
-  const action = useMemo(() => {
-    switch (provider as ModelProvider) {
-      case ModelProvider.Bedrock: {
-        return <BedrockForm />;
-      }
-
-      case ModelProvider.Google: {
-        return <GoogleForm />;
-      }
-
-      case ModelProvider.ZhiPu: {
-        return <ZhipuForm />;
-      }
-
-      case ModelProvider.Mistral: {
-        return <MistralForm />;
-      }
-
-      case ModelProvider.Moonshot: {
-        return <MoonshotForm />;
-      }
-
-      case ModelProvider.Perplexity: {
-        return <PerplexityForm />;
-      }
-
+  const apiKeyPlaceholder = useMemo(() => {
+    switch (provider) {
       case ModelProvider.Anthropic: {
-        return <AnthropicForm />;
-      }
-
-      case ModelProvider.Groq: {
-        return <GroqForm />;
+        return 'sk-ant_*****************************';
       }
 
       case ModelProvider.OpenRouter: {
-        return <OpenRouterForm />;
+        return 'sk-or-********************************';
       }
 
-      case ModelProvider.TogetherAI: {
-        return <TogetherAIForm />;
+      case ModelProvider.Perplexity: {
+        return 'pplx-********************************';
       }
 
-      case ModelProvider.ZeroOne: {
-        return <ZeroOneForm />;
+      case ModelProvider.ZhiPu: {
+        return '*********************.*************';
       }
 
-      default:
-      case ModelProvider.OpenAI: {
-        return <OpenAIForm />;
+      case ModelProvider.Groq: {
+        return 'gsk_*****************************';
+      }
+
+      default: {
+        return '*********************************';
       }
     }
   }, [provider]);
 
   return (
     <Center gap={16} style={{ maxWidth: 300 }}>
-      {action}
+      {provider === ModelProvider.Bedrock ? (
+        <BedrockForm />
+      ) : (
+        <ProviderApiKeyForm
+          apiKeyPlaceholder={apiKeyPlaceholder}
+          avatar={<ProviderAvatar provider={provider as ModelProvider} />}
+          provider={provider as GlobalLLMProviderKey}
+          showEndpoint={provider === ModelProvider.OpenAI}
+        />
+      )}
       <Flexbox gap={12} width={'100%'}>
         <Button
           block
