+++ src/locales/default/common.ts
@@ -86,23 +86,6 @@ export default {
     'zh-TW': '繁体中文',
   },
   layoutInitializing: '正在加载布局...',
-  modelProvider: {
-    anthropic: 'Anthropic',
-    azure: 'Azure',
-    bedrock: 'AWS Bedrock',
-    google: 'Google',
-    groq: 'Groq',
-    mistral: 'Mistral AI',
-    moonshot: 'Moonshot AI',
-    ollama: 'Ollama',
-    oneapi: 'One API',
-    openai: 'OpenAI',
-    openrouter: 'OpenRouter',
-    perplexity: 'Perplexity',
-    togetherai: 'TogetherAI',
-    zeroone: '01.AI 零一万物',
-    zhipu: '智谱AI',
-  },
   noDescription: '暂无描述',
   oauth: 'SSO 登录',
   ok: '确定',
