+++ src/locales/default/error.ts
@@ -108,59 +108,7 @@ export default {
     responseError: '服务请求失败，请检查配置或重试',
   },
   unlock: {
-    apikey: {
-      Anthropic: {
-        description: '输入你的 Anthropic API Key 即可开始会话。应用不会记录你的 API Key',
-        title: '使用自定义 Anthropic API Key',
-      },
-      Bedrock: {
-        customRegion: '自定义服务区域',
-        description:
-          '输入你的 AWS AccessKeyId / SecretAccessKey 即可开始会话。应用不会记录你的鉴权配置',
-        title: '使用自定义 Bedrock 鉴权信息',
-      },
-      Google: {
-        description: '输入你的 Google API Key 即可开始会话。应用不会记录你的 API Key',
-        title: '使用自定义 Google API Key',
-      },
-      Groq: {
-        description: '输入你的 Groq API Key 即可开始会话。应用不会记录你的 API Key',
-        title: '使用自定义 Groq API Key',
-      },
-      Mistral: {
-        description: '输入你的 Mistral AI API Key 即可开始会话。应用不会记录你的 API Key',
-        title: '使用自定义 Mistral AI API Key',
-      },
-      Moonshot: {
-        description: '输入你的 Moonshot AI API Key 即可开始会话。应用不会记录你的 API Key',
-        title: '使用自定义 Moonshot AI API Key',
-      },
-      OpenAI: {
-        addProxyUrl: '添加 OpenAI 代理地址（可选）',
-        description: '输入你的 OpenAI API Key 即可开始会话。应用不会记录你的 API Key',
-        title: '使用自定义 OpenAI API Key',
-      },
-      OpenRouter: {
-        description: '输入你的 OpenRouter API Key 即可开始会话。应用不会记录你的 API Key',
-        title: '使用自定义 OpenRouter API Key',
-      },
-      Perplexity: {
-        description: '输入你的 Perplexity API Key 即可开始会话。应用不会记录你的 API Key',
-        title: '使用自定义 Perplexity API Key',
-      },
-      TogetherAI: {
-        description: '输入你的 TogetherAI API Key 即可开始会话。应用不会记录你的 API Key',
-        title: '使用自定义 TogetherAI API Key',
-      },
-      ZeroOne: {
-        description: '输入你的零一万物 API Key 即可开始会话。应用不会记录你的 API Key',
-        title: '使用自定义零一万物 API Key',
-      },
-      Zhipu: {
-        description: '输入你的 Zhipu API Key 即可开始会话。应用不会记录你的 API Key',
-        title: '使用自定义 Zhipu API Key',
-      },
-    },
+    addProxyUrl: '添加 OpenAI 代理地址（可选）',
     closeMessage: '关闭提示',
     confirm: '确认并重试',
     model: {
